### 2021-04-16 #3176 @whedon
[PRE REVIEW]: pypbomb: A Python package with tools for the design of detonation tubes
**Submitting author:** @cartemic (<a href="http://orcid.org/0000-0002-5559-936X">Mick Carter</a>)
**Repository:** <a href="https://github.com/cartemic/pypbomb/" target ="_blank">https://github.com/cartemic/pypbomb/</a>
**Version:** v1.1.0
