### 2021-06-22 #3176 @kyleniemeyer
Hello @mefuller @arosen93, would you be willing to review this submission for JOSS?
