### 2021-06-22 #3176 @kyleniemeyer
Hi @fwitte, could you review this submission for JOSS?
