### 2021-06-29 #3176 @kyleniemeyer
@mefuller you can certainly begin looking at the software, based on our [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), but I need to identify a second reviewer before I start the actual review issue
