### 2021-09-30 #3176 @kyleniemeyer
Hello @mwerwee, are you able to review this submission for JOSS? It focuses on the design and analysis of detonation tubes.
