### 2022-02-09 #3176 @jiweiqi
hi @kyleniemeyer , thanks for your invitation! Detonation is out of my expertise. I would recommend @wenkailiang.
