### 2022-02-10 #4143 @whedon
[REVIEW]: pypbomb: A Python package with tools for the design of detonation tubes
**Submitting author:** @cartemic (<a href="http://orcid.org/0000-0002-5559-936X">Mick Carter</a>)
**Repository:** https://github.com/cartemic/pypbomb/
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
