### 2022-02-11 4268dc7
Add JOSS badge to README
