### 2022-05-04 cartemic/pypbomb#7 @mefuller
Please update and verify installation instructions to link to homepages for dependencies to provide alternate installation methods (e.g. cantera and pint both available with pip, some system package managers)
