### 2022-05-04 cartemic/pypbomb#8 @mefuller
JOSS review: example usage and built-ins in docs
In the example usage, no link is provided for a list of allowed materials nor is the source of their properties apparent

Please also make clear that the "sizes" entered are NPS (no units)

Can / How should the program be applied to tubing or piping that is not NPS/schedule, but defined instead with actual dimensions? I.e. is there an option to supply two of three of ID,OD, and wall thickness? What about metric?

Can one create custom materials/mechanical properties?

Links to relevant documentation/tables/etc. for NPS, ASME flanges, and UNC/UNF fasteners (andany other supported standard hardware) would be appreciated
