### 2022-05-04 #4143 @kyleniemeyer
@cartemic Can you address @mefuller's comments (left as issues https://github.com/cartemic/pypbomb/issues/7 and https://github.com/cartemic/pypbomb/issues/8) and let me know here when those are resolved?
