### 2022-05-05 cartemic/pypbomb#7 @cartemic
Oh neat, cantera is on pip now! That should simplify things a lot.
