### 2022-05-05 cartemic/pypbomb#7 @cartemic
This may take some time. The new version of cantera doesn't seem to work with my code.
