### 2022-05-05 cartemic/pypbomb#7 @cartemic
- [x] Change installer to pip only
- [ ] Adjust code for new version of Cantera
- [ ] Adjust CI to use new installation process
- [ ] Unify CI and install dependencies to single file
- [ ] Update documentation
