### 2022-05-05 4229859
Fix object casting in interpolation dataframe
I think this is from a new version of pandas
