### 2022-05-05 a78a57b
Rework installation and CI for strict pip install
