+++ requirements.txt
@@ -1,0 +1,5 @@
+cantera==2.6.0
+numpy==1.22.3
+pandas==1.4.2
+pint<=0.11
+scipy==1.8.0
