+++ .github/workflows/Documentation.yml
@@ -32,19 +32,7 @@ jobs:
     # why.
     - name: Run, Ronnie, Run
       run: |
-        wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
-        bash miniconda.sh -b -p $HOME/miniconda
-        rm miniconda.sh
-        source $HOME/miniconda/etc/profile.d/conda.sh && conda activate
-        conda config --set always_yes yes --set changeps1 no
-        conda update -q conda
-        conda update -q --all
-        conda config --append channels conda-forge
-        conda env create -qf test-environment.yaml
-        conda info --envs
-        source activate test-environment
-        conda info --envs
-        pip install wheel setuptools
+        pip install -r requirements.txt
         pip install -r docs/requirements.txt
         cd docs/
         make clean html
