### 2022-05-05 #4143 @wenkailiang
@kyleniemeyer @cartemic This is a nice contribution to the community. The package could be very useful for the design of detonation tube. Just a few comments:

- The method used in the DDT estimation is not very clear to me (it might be due to my lack of knowledge in this area). Some equations from the Ciccarelli and Dorofeev work might be added to the manuscript.

- Is it possible to include the non-ideal effects in the calculations? For example,  https://github.com/Cantera/cantera/blob/master/interfaces/cython/cantera/examples/reactors/NonIdealShockTube.py
