### 2022-05-06 40a556b
Update installation instructions with links to dependencies
