### 2022-05-06 cartemic/pypbomb#7 @cartemic
@mefuller I've added a PR that I hope will resolve this issue. If you'd like, I would be happy to add you as a contributor if you would like to approve the PR. Otherwise I will hold off on merging it to master until you're happy with the changes. The [updated documentation](https://cartemic.github.io/pypbomb/installation.html) is currently live. Please let me know what you think!
