### 2022-05-06 cartemic/pypbomb#10 @cartemic
Add DDT run-up equations to paper
@kyleniemeyer @cartemic This is a nice contribution to the community. The package could be very useful for the design of detonation tube. Just a few comments:

- The method used in the DDT estimation is not very clear to me (it might be due to my lack of knowledge in this area). Some equations from the Ciccarelli and Dorofeev work might be added to the manuscript.

- Is it possible to include the non-ideal effects in the calculations? For example,  https://github.com/Cantera/cantera/blob/master/interfaces/cython/cantera/examples/reactors/NonIdealShockTube.py

_Originally posted by @wenkailiang in https://github.com/openjournals/joss-reviews/issues/4143#issuecomment-1118196885_
