### 2022-05-06 #4143 @cartemic
@wenkailiang thanks for your comments! I've opened an issue to include equations for DDT run-up estimation in the paper (cartemic/pypbomb#10).

Non-ideal effects are not currently within the scope of pypbomb as presented here, however they could be included in a future version by exposing the option for phase specification to the user in more places (e.g. the `sd` module). Currently phase specification within pypbomb is only used in DDT run-up calculations, and the inclusion there is due to the use of viscosity in the run-up equation used for blockage ratios less than 30%.
