### 2022-05-07 cartemic/pypbomb#8 @cartemic
- [ ] Add list of available materials -- probably should be auto-generated with precommit script
- [ ] Find more info about properties, tables, etc.
- [ ] Add information on editing properties + materials
- [ ] Check on non-NPS piping

Not exactly part of the issue, but since I'm in here working on documentation and planning to do some precommit script work for the materials list, this seems like a good time to implement some general code-quality stuff that I think is important, but didn't have on my radar when I submitted this for publication.

- [ ] Add black, isort, flake8 with precommit hooks
- [ ] Add section on contributing to the README and documentation pages
