### 2022-05-07 b1d3495
Add code quality checks
