+++ pypbomb/thermochem.py
@@ -10,19 +10,18 @@ import cantera as ct
 import numpy as np
 import pint
 
-from . import units, sd
-
+from . import sd, units
 
 _U = pint.UnitRegistry()
 
 
 def calculate_laminar_flame_speed(
-        initial_temperature,
-        initial_pressure,
-        species,
-        mechanism,
-        phase_specification="",
-        unit_registry=_U
+    initial_temperature,
+    initial_pressure,
+    species,
+    mechanism,
+    phase_specification="",
+    unit_registry=_U,
 ):
     """
     This function uses cantera to calculate the laminar flame speed of a given
@@ -52,29 +51,15 @@ def calculate_laminar_flame_speed(
     gas = ct.Solution(mechanism, phase_specification)
     quant = unit_registry.Quantity
 
-    initial_pressure = units.parse_quant_input(
-        initial_pressure,
-        unit_registry
-    )
-    initial_temperature = units.parse_quant_input(
-        initial_temperature,
-        unit_registry
-    )
-    units.check_pint_quantity(
-        initial_pressure,
-        "pressure",
-        ensure_positive=True
-    )
-    units.check_pint_quantity(
-        initial_temperature,
-        "temperature",
-        ensure_positive=True
-    )
+    initial_pressure = units.parse_quant_input(initial_pressure, unit_registry)
+    initial_temperature = units.parse_quant_input(initial_temperature, unit_registry)
+    units.check_pint_quantity(initial_pressure, "pressure", ensure_positive=True)
+    units.check_pint_quantity(initial_temperature, "temperature", ensure_positive=True)
 
     gas.TPX = (
         initial_temperature.to("K").magnitude,
         initial_pressure.to("Pa").magnitude,
-        species
+        species,
     )
 
     # find laminar flame speed
@@ -86,14 +71,7 @@ def calculate_laminar_flame_speed(
 
 
 # noinspection SpellCheckingInspection
-def get_eq_sound_speed(
-        temperature,
-        pressure,
-        species,
-        mechanism,
-        phase_specification="",
-        unit_registry=_U
-):
+def get_eq_sound_speed(temperature, pressure, species, mechanism, phase_specification="", unit_registry=_U):
     """
     Calculates the equilibrium speed of sound in a mixture
 
@@ -120,31 +98,17 @@ def get_eq_sound_speed(
     """
     quant = unit_registry.Quantity
 
-    pressure = units.parse_quant_input(
-        pressure,
-        unit_registry
-    )
-    temperature = units.parse_quant_input(
-        temperature,
-        unit_registry
-    )
-    units.check_pint_quantity(
-        pressure,
-        "pressure",
-        ensure_positive=True
-    )
-    units.check_pint_quantity(
-        temperature,
-        "temperature",
-        ensure_positive=True
-    )
+    pressure = units.parse_quant_input(pressure, unit_registry)
+    temperature = units.parse_quant_input(temperature, unit_registry)
+    units.check_pint_quantity(pressure, "pressure", ensure_positive=True)
+    units.check_pint_quantity(temperature, "temperature", ensure_positive=True)
 
     working_gas = ct.Solution(mechanism, phase_specification)
     working_gas.TPX = [
         temperature.to("K").magnitude,
         pressure.to("Pa").magnitude,
-        species
-        ]
+        species,
+    ]
 
     pressures = np.zeros(2)
     densities = np.zeros(2)
@@ -161,18 +125,18 @@ def get_eq_sound_speed(
     densities[1] = working_gas.density
 
     # calculate sound speed
-    sound_speed = np.sqrt(np.diff(pressures)/np.diff(densities))[0]
[…]
