+++ pypbomb/tests/test_tube.py
@@ -9,8 +9,7 @@ import pandas as pd
 import pytest
 from mock import patch
 
-from .. import thermochem
-from .. import tube
+from .. import thermochem, tube
 
 
 def compare(manual, tested):
@@ -56,7 +55,7 @@ class TestBolt:
             bolt_max_tensile,
             plate_max_tensile,
             engagement_length,
-            tube._U
+            tube._U,
         )
 
         compare(hand_calc, test_values)
@@ -85,7 +84,7 @@ class TestBolt:
             bolt_max_tensile,
             plate_max_tensile,
             engagement_length,
-            tube._U
+            tube._U,
         )
 
         compare(hand_calc, test_values)
@@ -96,7 +95,7 @@ class TestBolt:
         hand_calc = {
             "screw area": tube._Q(0.034934049, "in^2"),
             "plate area": tube._Q(0.308744082, "in^2"),
-            "minimum engagement": tube._Q(0.452595544, "in")
+            "minimum engagement": tube._Q(0.452595544, "in"),
         }
         test_areas = tube.Bolt.calculate_stress_areas(
             self.thread_size,
@@ -104,7 +103,7 @@ class TestBolt:
             bolt_max_tensile,
             self.plate_max_tensile,
             self.engagement_length,
-            tube._U
+            tube._U,
         )
         compare(hand_calc, test_areas)
 
@@ -113,7 +112,7 @@ class TestBolt:
         bolt_max_tensile = tube._Q(80, "ksi")
         hand_calc = {
             "screw area": tube._Q(0.036374073, "in^2"),
-            "minimum engagement": tube._Q(0.314168053, "in")
+            "minimum engagement": tube._Q(0.314168053, "in"),
         }
         test_areas = tube.Bolt.calculate_stress_areas(
             self.thread_size,
@@ -121,7 +120,7 @@ class TestBolt:
             bolt_max_tensile,
             self.plate_max_tensile,
             self.engagement_length,
-            tube._U
+            tube._U,
         )
         compare(hand_calc, test_areas)
 
@@ -129,10 +128,10 @@ class TestBolt:
         # ensure warning when engagement length < minimum
         engagement_length = tube._Q(0.05, "in")
         with pytest.warns(
-                Warning,
-                match="Screws fail in shear, not tension." +
-                      " Plate may be damaged." +
-                      " Consider increasing bolt engagement length"
+            Warning,
+            match="Screws fail in shear, not tension."
+            + " Plate may be damaged."
+            + " Consider increasing bolt engagement length",
         ):
             tube.Bolt.calculate_stress_areas(
                 self.thread_size,
@@ -140,30 +139,17 @@ class TestBolt:
                 self.bolt_max_tensile,
                 self.plate_max_tensile,
                 engagement_length,
-                tube._U
+                tube._U,
             )
 
     @staticmethod
     def test_get_thread_property():
         # good input
         good_args = [
-            [
-                "pitch diameter max",
-                "1/4-20",
-                "2B",
-                tube._U
-            ],
-            [
-                "pitch diameter max",
-                "1/4-20",
-                "2A",
-                tube._U
-            ]
+            ["pitch diameter max", "1/4-20", "2B", tube._U],
+            ["pitch diameter max", "1/4-20", "2A", tube._U],
         ]
-        good_results = [
-            0.2248,
-            0.2164
-        ]
+        good_results = [0.2248, 0.2164]
         for args, result in zip(good_args, good_results):
             test_result = tube.Bolt.get_thread_property(*args)
 
@@ -174,63 +160,32 @@ class TestBolt:
         # property not in dataframe
         bad_property = "jello"
         bad_message = (
[…]
