+++ docs/conf.py
@@ -13,15 +13,16 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../'))
-from pypbomb import __version__ as release
 
+from pypbomb import __version__ as release  # noqa:F401,E402
+
+sys.path.insert(0, os.path.abspath("../"))
 
 # -- Project information -----------------------------------------------------
 
-project = 'pypbomb'
-copyright = '2020, Mick Carter'
-author = 'Mick Carter'
+project = "pypbomb"
+copyright = "2020, Mick Carter"
+author = "Mick Carter"
 
 # # The full version, including alpha/beta/rc tags
 # release = '0.1.0'
@@ -36,19 +37,19 @@ extensions = [
     "sphinx.ext.autosummary",
     "sphinx.ext.napoleon",
     "sphinx_rtd_theme",
-    "nbsphinx"
+    "nbsphinx",
 ]
 
 autosummary_generate = True
 napoleon_google_docstring = False
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -61,7 +62,7 @@ html_theme = "sphinx_rtd_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 
 def setup(app):

[…]
