### 2022-05-11 072ec52
Remove commas after equations. They look silly when TeX'd.
