### 2022-05-11 cartemic/pypbomb#10 @cartemic
@wenkailiang when you get a chance please let me know if this addresses your concern adequately and I'll get it merged in.
