### 2022-05-11 #4143 @cartemic
@kyleniemeyer [#10](https://github.com/cartemic/pypbomb/issues/10) is resolved.
