### 2022-05-12 0918bf8
Update docs/requirements.txt to include reqs for example notebook
