### 2022-05-12 043f77d
Quiet some unnecessary warnings
