### 2022-05-12 8b74696
Add available_materials to base docstrings for Tube, Flange
