### 2022-05-13 #4143 @cartemic
@kyleniemeyer [#7](https://github.com/cartemic/pypbomb/issues/7) is resolved.
