### 2022-05-13 2cf2a7c
Update materials list and associated tests
