### 2022-05-13 cartemic/pypbomb#8 @cartemic
@mefuller I have added PR #14, which I hope adequately addresses the concerns raised in this issue. When you have a chance, please let me know if this is sufficient, or if there is anything else you would like me to add or clarify. Thank you!
