+++ _sources/index.rst.txt
@@ -4,7 +4,16 @@
    contain the root `toctree` directive.
 
 Pypbomb
-================================================================
+=======
+
+.. image:: https://camo.githubusercontent.com/d91ed7ac7abbd5a6102cbe988dd8e9ac21bde0a73d97be7603b891ad08ce3479/68747470733a2f2f696d672e736869656c64732e696f2f62616467652f636f64652532307374796c652d626c61636b2d3030303030302e737667
+   :alt: Black badge
+   :target: https://github.com/psf/black
+
+.. image:: https://camo.githubusercontent.com/fe4a658dd745f746410f961ae45d44355db1cc0e4c09c7877d265c1380248943/68747470733a2f2f696d672e736869656c64732e696f2f62616467652f253230696d706f7274732d69736f72742d2532333136373462313f7374796c653d666c6174266c6162656c436f6c6f723d656638333336
+   :alt: Isort badge
+   :target: https://github.com/PyCQA/isort
+
 .. image:: https://api.travis-ci.com/cartemic/pypbomb.svg?branch=master
    :alt: Travis build badge
    :target: https://travis-ci.com/github/cartemic/pypbomb
@@ -17,6 +26,11 @@ Pypbomb
    :alt: Documentation badge
    :target: https://cartemic.github.io/pypbomb/
 
+.. image:: https://joss.theoj.org/papers/288143539a2e4f5ab1b83cbee4e17052/status.svg
+   :alt: JOSS badge
+   :target: https://joss.theoj.org/papers/288143539a2e4f5ab1b83cbee4e17052
+
+
 For when you want the detonations to stay on the inside.
 
 This package was made to help with the design of a closed-end detonation tube,
@@ -27,10 +41,10 @@ at Oregon Sate University.
 Project files can be found `here <https://github.com/cartemic/pypbomb>`_.
 
 .. toctree::
-   :maxdepth: 2
-   :caption: Installation and Testing
+   :maxdepth: 1
+   :caption: Getting Started
 
-   installation
+   getting_started
 
 .. toctree::
    :maxdepth: 1
