### 2022-05-13 #4143 @kyleniemeyer
Hi @cartemic, we are nearly ready to accept! Just a few small items:

1. Can you add commas after the uses of "e.g." in the paper? In other words, "e.g.,".

2. Could you add a bit more detail to the README? Since that is the effectively the landing page for people, I think it should have a bit more basic info about the project, simple installation instructions (not for developers, just pip/conda users), and perhaps a very simple example usage.

Once you have done these things, please archive the software repository on Zenodo or Figshare and let me know the DOI here.
