+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 A detonation is supersonic combustion in which a reaction front is coupled with a pressure shock front [@Lee2008]. Adiabatic heating from the shock front helps to sustain the combustion, which in turn accelerates the shock front and allows it to propagate supersonically. Because the products of a detonation are at a higher pressure than the reactants, thermodynamic cycles using detonations, such as the Humphrey cycle, have the potential for higher thermodynamic efficiency than deflagration-based cycles, such as the Brayton cycle [@Coleman2001]. However, knowledge of a reactant mixture's characteristic detonation structure is needed in order to make practical use of it with detonation-based cycles.
 
-The behavior of a detonation can be investigated either via simulation or physical experimentation. Given that detonation is an inherently three-dimensional process [@Lee2008; @ciccarelli], accurate simulations can be computationally expensive [@Kessler2010], requiring runs on the order of days to months for a single 2-D simulation depending on the hardware involved [@Kessler2011; @Radulescu2007]. Alternatively, a detonation tube is often used to experimentally study the behavior of detonations. In a detonation tube, data for a given set of initial conditions can be collected anywhere from multiple times per second (e.g. in a pulse detonation engine) to multiple times per hour (e.g. in a closed-end detonation tube).
+The behavior of a detonation can be investigated either via simulation or physical experimentation. Given that detonation is an inherently three-dimensional process [@Lee2008; @ciccarelli], accurate simulations can be computationally expensive [@Kessler2010], requiring runs on the order of days to months for a single 2-D simulation depending on the hardware involved [@Kessler2011; @Radulescu2007]. Alternatively, a detonation tube is often used to experimentally study the behavior of detonations. In a detonation tube, data for a given set of initial conditions can be collected anywhere from multiple times per second (e.g., in a pulse detonation engine) to multiple times per hour (e.g., in a closed-end detonation tube).
 
 The design of a detonation tube requires many considerations, including estimation of the required length for deflagration-to-detonation transition (DDT), tube material and size selection (including the effects of transient pressures), fastener failure calculations (including bolt pull-out), flange class selection, viewing window sizing (if optical access is required), and prediction of safe operating conditions (including accounting for detonation reflection in the case of a closed-end tube). All of this is specific to the mixture being detonated, therefore it is important to be able to quickly re-perform the analysis for new reactant mixtures. The tools within ``pypbomb`` provide a first-order analysis meant to serve as the basis for the previously detailed analysis.
 
