### 2022-05-17 e193278
Add more detail to README
