+++ README.md
@@ -6,6 +6,7 @@
 [![Codecov badge](https://codecov.io/gh/cartemic/pypbomb/branch/master/graph/badge.svg)](https://codecov.io/gh/cartemic/pypbomb)
 [![Documentation badge](https://github.com/cartemic/pypbomb/workflows/Documentation/badge.svg)](https://cartemic.github.io/pypbomb/)
 [![status](https://joss.theoj.org/papers/288143539a2e4f5ab1b83cbee4e17052/status.svg)](https://joss.theoj.org/papers/288143539a2e4f5ab1b83cbee4e17052)
+[![DOI](https://zenodo.org/badge/130134402.svg)](https://zenodo.org/badge/latestdoi/130134402)
 
 ---
 
