### 2022-05-18 3d5b660
Update Documentation.yml
Allow manual trigger

[no ci]
