+++ .github/workflows/Documentation.yml
@@ -9,6 +9,7 @@ on:
     branches: [ master, ppf_sphinx ]
   pull_request:
     branches: [ master ]
+  workflow_dispatch:
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
