+++ _sources/index.rst.txt
@@ -6,11 +6,11 @@
 Pypbomb
 =======
 
-.. image:: https://camo.githubusercontent.com/d91ed7ac7abbd5a6102cbe988dd8e9ac21bde0a73d97be7603b891ad08ce3479/68747470733a2f2f696d672e736869656c64732e696f2f62616467652f636f64652532307374796c652d626c61636b2d3030303030302e737667
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
    :alt: Black badge
    :target: https://github.com/psf/black
 
-.. image:: https://camo.githubusercontent.com/fe4a658dd745f746410f961ae45d44355db1cc0e4c09c7877d265c1380248943/68747470733a2f2f696d672e736869656c64732e696f2f62616467652f253230696d706f7274732d69736f72742d2532333136373462313f7374796c653d666c6174266c6162656c436f6c6f723d656638333336
+.. image:: https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat
    :alt: Isort badge
    :target: https://github.com/PyCQA/isort
 
