### 2022-05-18 #4143 @kyleniemeyer
Congratulations @cartemic on your paper's publication in JOSS!

Many thanks to @mefuller and @wenkailiang for reviewing this submission.
