### 2022-05-23 #4143 @JohannesBuchner
I found this because I am subscribed to the JOSS news feed. To an unfamiliar person such as myself, this software seems to be applicable to improving construction of pipe bombs and to estimate the impact of attacks on infrastructure.  I am a bit puzzled why neither the review proces nor the paper includes a discussion of ethical concerns of a public release of this software? For example, an argument why this is not a concern to the authors? Maybe I am misunderstanding something. From looking at the research group's profile, they seem to be involved in military applications ("projectile impact").
